2.0.161.3

+++ @vue-modules/vue/index/index.js
@@ -4,11 +4,16 @@
 
 const index_dir = ß.VAR + "/app";
 
-ß.fs.removeSync(index_dir);
-ß.fs.mkdirpSync(index_dir);
-
-ß.fs.writeFileSync(index_dir + '/index.html', '<head><META HTTP-EQUIV="refresh" CONTENT="1"></head><body><input type="button" value = "Vue build index .." onclick="history.go(0)" /></body>');
-
+if (process.argv.indexOf("--restart-server") >= 0) ß.ntc("Skipping forks, restart-server");
+else {
+    // a clean restart, wipe existing content
+    ß.fs.removeSync(index_dir);
+    ß.fs.mkdirpSync(index_dir);
+    ß.fs.writeFileSync(
+        index_dir + "/index.html",
+        '<head><META HTTP-EQUIV="refresh" CONTENT="1"></head><body><input type="button" value = "Vue build index .." onclick="history.go(0)" /></body>'
+    );
+}
 // the express index dir will contain our possibly multilingual language pages
 ß.app.use(ß.express.static(index_dir, ß.STATIC_OPTIONS));
